Remove matchProtocol option. | Rework matching orders. | Add support for page/site specific lists. | Minor GUI tweaks.
