+++ pages/options/options.html
@@ -13,7 +13,7 @@
   
   <!-- Page resources -->
   <link rel="stylesheet" type="text/css" href="./options.css"/>
-  <script type="text/javascript" src="./options-new.js"></script>
+  <script type="text/javascript" src="./options.js"></script>
 </head>
 <body>
   <h2 class="header">Backloader Filter List</h2>
